\name{hlap-package}
\alias{hlap-package}
\alias{hlap}
\docType{package}
\title{
\packageTitle{hlap}
}
\description{
\packageDescription{hlap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hlap}
\packageIndices{hlap}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hlap}

Maintainer: \packageMaintainer{hlap}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
